update 19-9-2020: add AccessDenied, Authorization

+++ public/controllers/DeleteController.php
@@ -2,16 +2,20 @@
 require_once '../../admin/controllers/UserController.php';
 require_once "../config/routes.php";
 
-if ($_SERVER["REQUEST_METHOD"] == "GET") {
-    $userId = $_GET["userid"];
+session_start();
 
-    //echo $userId;
-    $controller = new UserController();
-    $rs = $controller->DeleteUser($userId);
-    if($rs){
-        header("location: ".ROUTE_HOME);
-    }
-    else {
+if ($_SESSION["role"] == 1) {
+    header("location: " . ROUTE_ACCESSDENIED);
+} else {
+    if ($_SERVER["REQUEST_METHOD"] == "GET") {
+        $userId = $_GET["userid"];
+
+        $controller = new UserController();
+        $rs = $controller->DeleteUser($userId);
         
+        if ($rs) {
+            header("location: " . ROUTE_HOME);
+        } else {
+        }
     }
 }
